feat: split Logs-Builder into Basic and Advanced shards
Split logsQueryBuilder-filters into basic (13 tests) and advanced (13 tests
+ 2 SQL OFF) files. Replace single Logs-Builder shard with Logs-Builder-Basic
(chart + filters-basic + pagination + unflattened) and Logs-Builder-Advanced
(editor + filters-advanced + quickmode + histogram) to cut shard time.
